Asset Manager/requirements.txt
# requirements.txt
Pillow>=8.0.0
numpy>=1.20.0